Moves FormFieldTextEditor
